+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Hide/Flatten all buildings, if searchPattern can't find any matching nodes #225
 - Show maxValue of each metric in metricChooser select list #204
 - Colored color-slider inside the RibbonBar #318
+- Color positive buildings white and set as default option #311
 - Adapt colorRange in when changing colorMetric #330
 
 ### Changed